admxgen: split tests to internal and package testing
As we now have an API, separate the tests exercising the API.
